migrate accordion/toggler to testing-library
